# Speed up type literal lexing and make it more strict.