--- a/toolchain/lex/lex.cpp
+++ b/toolchain/lex/lex.cpp
@@ -1188,7 +1188,13 @@ auto Lexer::LexWordAsTypeLiteralToken(llvm::StringRef word, int32_t byte_offset)
   }
 
   llvm::StringRef suffix = word.substr(1);
-  constexpr ssize_t DigitLimit = std::numeric_limits<int64_t>::digits10;
+
+  // Type bit-widths can't usefully be large integers so we restrict to small
+  // ones that are especially easy to parse into a normal integer variable by
+  // restricting the number of digits to round trip.
+  int64_t suffix_value;
+  constexpr ssize_t DigitLimit =
+      std::numeric_limits<decltype(suffix_value)>::digits10;
   if (suffix.size() > DigitLimit) {
     // See if this is not actually a type literal.
     if (!llvm::all_of(suffix, IsDecimalDigit)) {
@@ -1208,12 +1214,10 @@ auto Lexer::LexWordAsTypeLiteralToken(llvm::StringRef word, int32_t byte_offset)
   // It's tempting to do something more clever because we know the length ahead
   // of time, but we expect these to be short (1-3 digits) and profiling doesn't
   // show the loop as hot in the short cases.
-  ssize_t suffix_value = suffix[0] - '0';
-  for (char c : suffix.drop_front()) {
-    if (!IsDecimalDigit(c)) {
-      return LexResult::NoMatch();
-    }
-    suffix_value = suffix_value * 10 + (c - '0');
+  if (auto [ptr, error] =
+          std::from_chars(suffix.begin(), suffix.end(), suffix_value);
+      error != std::errc() || ptr != suffix.end()) {
+    return LexResult::NoMatch();
   }
 
   return LexTokenWithPayload(