@chandlerc 2024-10-24
> Approving, because I expect you're going to be of a different mind about the importance of sharing of code.

Thanks, will look at the other things in follow-ups!