@chandlerc 2024-10-24
I've restored the for loop for now, but maybe there is another approach? Nothing really simple comes to mind, and the loop is at least *somewhat* simple...

FWIW, it appears to be a bit faster to use the loop on x86. (About the same on Arm for me.)